moving all utils to folder in lib + readmes #92
