+++ lib/README.md
@@ -1,0 +1,3 @@
+# Lib <img src="../command/frontend/res/logo.png" alt="logo" width="20"/> 
+
+The lib is the library 📖 of tools that other servers use.
