+++ lib/index.js
@@ -1,14 +1,14 @@
 module.exports = {
-	webhookAlert: require("./webhookAlert.js"),
-	auth: require("./auth.js"),
-	handleServerStart: require("./handleServerStart.js"),
-	handleSecureServerStart: require("./handleSecureServerStart.js"),
-	validateBody: require("./validateBody.js"),
-	subprocess: require("./subprocess.js"),
-	formatBytes: require("./formatBytes.js"),
-	tempMiddleware: require("./tempMiddleware.js"),
-	jsonFileHanding: require("./jsonFileHandling.js"),
-	randomID: require("./randomID.js"),
-	helmetOptions: require("./helmetOptions.js"),
+	webhookAlert: require("./utils/webhookAlert.js"),
+	auth: require("./utils/auth.js"),
+	handleServerStart: require("./utils/handleServerStart.js"),
+	handleSecureServerStart: require("./utils/handleSecureServerStart.js"),
+	validateBody: require("./utils/validateBody.js"),
+	subprocess: require("./utils/subprocess.js"),
+	formatBytes: require("./utils/formatBytes.js"),
+	tempMiddleware: require("./utils/tempMiddleware.js"),
+	jsonFileHanding: require("./utils/jsonFileHandling.js"),
+	randomID: require("./utils/randomID.js"),
+	helmetOptions: require("./utils/helmetOptions.js"),
 	isPrimeInstance: !("NODE_APP_INSTANCE" in process.env) || process.env.NODE_APP_INSTANCE === "0"
 }
