+++ lib/module.js
@@ -1,1 +1,1 @@
-export var formatBytes = require("./formatBytes.js")
+export var formatBytes = require("./utils/formatBytes.js")
