+++ lib/utils/auth.js
