+++ lib/utils/formatBytes.js
