+++ lib/utils/handleSecureServerStart.js
