+++ lib/utils/helmetOptions.js
