+++ memory/README.md
@@ -1,0 +1,3 @@
+# Memory <img src="../command/frontend/res/logo.png" alt="logo" width="20"/> 
+
+The memory server is a socket that keeps track of events and callbacks for the other servers to call.
