+++ memory/lib/cron.js
@@ -1,0 +1,6 @@
+const cron = require("node-cron")
+
+module.exports = (cronString, cronTask) => { 
+    const fileStatsCron = cron.schedule(cronString, cronTask)
+    fileStatsCron.start()
+}
